FinancialAssistantAgent/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# ── LLM ────────────────────────────────────────────────────────────────────────
# On Windows, install separately with: pip install llama-cpp-python --prefer-binary
llama-cpp-python>=0.2.57

# ── RAG / Embeddings ───────────────────────────────────────────────────────────
# torch and transformers are pulled in automatically by sentence-transformers
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
numpy>=1.26.0

# ── Document parsing ───────────────────────────────────────────────────────────
PyPDF2>=3.0.0
python-docx>=1.1.0

# ── UI ─────────────────────────────────────────────────────────────────────────
gradio>=4.26.0

# ── HTTP client ────────────────────────────────────────────────────────────────
requests>=2.31.0

# ── Model download ─────────────────────────────────────────────────────────────
huggingface-hub>=0.22.0

# ── Utilities ──────────────────────────────────────────────────────────────────
# Note: sqlite3 is built into Python — never add it here
python-multipart>=0.0.9
httpx>=0.27.0

# ── Testing ────────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
